+++ chains/evm/listener/depositHandlers/erc20_test.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"math/big"
 	"testing"
+	"time"
 
 	"github.com/ChainSafe/sygma-relayer/e2e/evm"
 	"github.com/ChainSafe/sygma-relayer/relayer/transfer"
@@ -32,8 +33,15 @@ func TestRunErc20HandlerTestSuite(t *testing.T) {
 func (s *Erc20HandlerTestSuite) TestErc20HandleEvent() {
 	// 0xf1e58fb17704c2da8479a533f9fad4ad0993ca6b
 	recipientByteSlice := []byte{241, 229, 143, 177, 119, 4, 194, 218, 132, 121, 165, 51, 249, 250, 212, 173, 9, 147, 202, 107}
+	maxFee := big.NewInt(200000)
+	optionalMessage := common.LeftPadBytes(maxFee.Bytes(), 32)
+	optionalMessage = append(optionalMessage, []byte("optionalMessage")...)
+
+	metadata := make(map[string]interface{})
+	metadata["gasLimit"] = uint64(200000)
 
 	calldata := evm.ConstructErc20DepositData(recipientByteSlice, big.NewInt(2))
+	calldata = append(calldata, optionalMessage...)
 	depositLog := &events.Deposit{
 		DestinationDomainID: 0,
 		ResourceID:          [32]byte{0},
@@ -45,8 +53,9 @@ func (s *Erc20HandlerTestSuite) TestErc20HandleEvent() {
 
 	sourceID := uint8(1)
 	amountParsed := calldata[:32]
-	recipientAddressParsed := calldata[64:]
+	recipientAddressParsed := calldata[64 : 64+len(recipientByteSlice)]
 
+	timestamp := time.Now()
 	expected := &message.Message{
 		Source:      sourceID,
 		Destination: depositLog.DestinationDomainID,
@@ -56,11 +65,14 @@ func (s *Erc20HandlerTestSuite) TestErc20HandleEvent() {
 			Payload: []interface{}{
 				amountParsed,
 				recipientAddressParsed,
+				optionalMessage,
 			},
-			Type: transfer.FungibleTransfer,
+			Type:     transfer.FungibleTransfer,
+			Metadata: metadata,
 		},
-		Type: transfer.TransferMessageType,
-		ID:   "messageID",
+		Type:      transfer.TransferMessageType,
+		ID:        "messageID",
+		Timestamp: timestamp,
 	}
 	erc20DepositHandler := depositHandlers.Erc20DepositHandler{}
 	message, err := erc20DepositHandler.HandleDeposit(
@@ -70,7 +82,9 @@ func (s *Erc20HandlerTestSuite) TestErc20HandleEvent() {
 		depositLog.ResourceID,
 		depositLog.Data,
 		depositLog.HandlerResponse,
-		"messageID")
+		"messageID",
+		timestamp,
+	)
 
 	s.Nil(err)
 	s.NotNil(message)
@@ -104,6 +118,7 @@ func (s *Erc20HandlerTestSuite) TestErc20HandleEventIncorrectDataLen() {
 		depositLog.Data,
 		depositLog.HandlerResponse,
 		"messageID",
+		time.Now(),
 	)
 
 	s.Nil(message)