+++ comm/p2p/host.go
@@ -48,6 +48,7 @@ func NewHost(privKey crypto.PrivKey, networkTopology *topology.NetworkTopology, 
 // LoadPeers clears out peerstore and loads new peers into it
 func LoadPeers(h host.Host, peers []*peer.AddrInfo) {
 	for _, p := range h.Peerstore().Peers() {
+		h.Peerstore().RemovePeer(p)
 		h.Peerstore().ClearAddrs(p)
 	}